requirements/base.txt
autobahn==0.17.2
Jinja2==2.9.5
libsass==0.12.3
PyYAML==3.12
Twisted==17.1.0
